requirements.txt
requests>=2.31
beautifulsoup4>=4.12
httpx>=0.27
lxml>=5.1
python-dotenv>=1.0
aiofiles>=23.0
